requirements.txt
pandas
requests
psycopg2-binary
SQLAlchemy
python-dotenv
schedule